requirements.txt
torch
transformers
sentencepiece
h5py
nltk
einops
flash-attn
setuptools
datasets
hydra-core
lightning
tensorboard
standard-imghdr # Temporary fix the tensorboard error
lion-pytorch
tensordict
accelerate
wandb
zstandard # This is for the pile dataset from huggingface
calflops
hkkang_utils
triton